+++ test/integration/server_usage/post_process.py
@@ -9,7 +9,6 @@
         test/integration/server_usage/post_process.py
 
     Arguments:
-        None
 
 """
 
@@ -35,7 +34,6 @@ import mongo_lib.mongo_libs as mongo_libs
 import mongo_lib.mongo_class as mongo_class
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -44,10 +42,6 @@ class UnitTest(unittest.TestCase):
     """Class:  UnitTest
 
     Description:  Class which is a representation of a unit testing.
-
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:  None
 
     Methods:
         setUp -> Unit testing initilization.
@@ -65,7 +59,6 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
@@ -76,9 +69,10 @@ class UnitTest(unittest.TestCase):
         self.config_path = os.path.join(self.test_path, "config")
         self.cfg = gen_libs.load_module("configuration", self.config_path)
 
-        svr = mongo_class.Server(self.cfg.name, self.cfg.user, self.cfg.passwd,
-                                 self.cfg.host, self.cfg.port, self.cfg.auth,
-                                 self.cfg.conf_file)
+        svr = mongo_class.Server(
+            self.cfg.name, self.cfg.user, self.cfg.passwd, host=self.cfg.host,
+            port=self.cfg.port, auth=self.cfg.auth,
+            conf_file=self.cfg.conf_file)
         svr.connect()
 
         if self.cfg.db in svr.fetch_dbs():
@@ -96,7 +90,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with printing unformatted data.
 
         Arguments:
-            None
 
         """
 
@@ -113,7 +106,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test printing formatted data.
 
         Arguments:
-            None
 
         """
 
@@ -128,7 +120,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test inserting data into Mongo database.
 
         Arguments:
-            None
 
         """
 
@@ -139,13 +130,7 @@ class UnitTest(unittest.TestCase):
         coll = mongo_libs.crt_coll_inst(self.cfg, self.cfg.db, self.cfg.coll)
         coll.connect()
 
-        if coll.coll_cnt() == 1:
-            status = True
-
-        else:
-            status = False
-
-        self.assertTrue(status)
+        self.assertTrue(coll.coll_cnt() == 1)
 
     def tearDown(self):
 
@@ -154,17 +139,16 @@ class UnitTest(unittest.TestCase):
         Description:  Clean up of integration testing.
 
         Arguments:
-            None
 
         """
 
-        db = mongo_class.DB(self.cfg.name, self.cfg.user, self.cfg.passwd,
-                            self.cfg.host, self.cfg.port, self.cfg.db,
-                            self.cfg.auth, self.cfg.conf_file)
-
-        db.db_connect(self.cfg.db)
-        db.db_cmd("dropDatabase")
-        cmds_gen.disconnect([db])
+        mongo = mongo_class.DB(
+            self.cfg.name, self.cfg.user, self.cfg.passwd, host=self.cfg.host,
+            port=self.cfg.port, db=self.cfg.db, auth=self.cfg.auth,
+            conf_file=self.cfg.conf_file)
+        mongo.db_connect(self.cfg.db)
+        mongo.db_cmd("dropDatabase")
+        cmds_gen.disconnect([mongo])
 
 
 if __name__ == "__main__":