+++ test/unit/server_usage/main.py
@@ -9,7 +9,6 @@
         test/unit/server_usage/main.py
 
     Arguments:
-        None
 
 """
 
@@ -33,7 +32,6 @@ import server_usage
 import lib.gen_libs as gen_libs
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -43,12 +41,9 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:  None
-
     Methods:
         setUp -> Unit testing initilization.
+        test_run_program -> Test with run_program call.
         test_dir_chk_crt_false -> Test with arg_dir_chk_crt returns False.
         test_dir_chk_crt_true -> Test with arg_dir_chk_crt returns True.
         test_require_false -> Test with arg_require returns False.
@@ -67,11 +62,31 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
         self.args = {"-c": "config_file", "-d": "config_dir"}
+
+    @mock.patch("server_usage.run_program")
+    @mock.patch("server_usage.arg_parser")
+    @mock.patch("server_usage.gen_libs")
+    def test_run_program(self, mock_lib, mock_arg, mock_run):
+
+        """Function:  test_run_program
+
+        Description:  Test with run_program call.
+
+        Arguments:
+
+        """
+
+        mock_lib.help_func.return_value = False
+        mock_lib.root_run.return_value = True
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_run.return_value = True
+
+        self.assertFalse(server_usage.main())
 
     @mock.patch("server_usage.run_program")
     @mock.patch("server_usage.arg_parser")
@@ -83,7 +98,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with arg_dir_chk_crt returns False.
 
         Arguments:
-            None
 
         """
 
@@ -104,7 +118,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with arg_dir_chk_crt returns True.
 
         Arguments:
-            None
 
         """
 
@@ -124,7 +137,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with arg_require returns False.
 
         Arguments:
-            None
 
         """
 
@@ -144,7 +156,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with arg_require returns True.
 
         Arguments:
-            None
 
         """
 
@@ -163,7 +174,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with root_run returns True.
 
         Arguments:
-            None
 
         """
 
@@ -181,7 +191,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with root_run returns False.
 
         Arguments:
-            None
 
         """
 
@@ -199,7 +208,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with help_func returns False.
 
         Arguments:
-            None
 
         """
 
@@ -218,7 +226,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test with help_func returns True.
 
         Arguments:
-            None
 
         """