+++ test/unit/server_usage/run_program.py
@@ -9,7 +9,6 @@
         test/unit/server_usage/run_program.py
 
     Arguments:
-        None
 
 """
 
@@ -25,8 +24,8 @@ else:
     import unittest
 
 # Third-party
-import mock
 import collections
+import mock
 
 # Local
 sys.path.append(os.getcwd())
@@ -41,44 +40,27 @@ class ProgramLock(object):
 
     """Class:  ProgramLock
 
-    Description:  Mock of the gen_class.ProgramLock class.
-
-    Super-Class:  object
-
-    Sub-Classes:
+    Description:  Class stub holder for gen_class.ProgramLock class.
 
     Methods:
-        __init__ -> Class instance initilization.
-        __del__ -> Deletion of the ProgramLock instance.
+        __init__ -> Class initialization.
 
     """
 
-    def __init__(self, argv, flavor_id=""):
+    def __init__(self, cmdline, flavor):
 
         """Method:  __init__
 
-        Description:  Initialization of an instance of the ProgramLock class.
+        Description:  Class initialization.
 
         Arguments:
-            (input) argv -> Arguments from the command line.
-            (input) flavor_id -> Unique identifier for an instance.
+            (input) cmdline -> Argv command line.
+            (input) flavor -> Lock flavor ID.
 
         """
 
-        self.lock_created = True
-
-    def __del__(self):
-
-        """Method:  __del__
-
-        Description:  Deletion of the ProgramLock instance.
-
-        Arguments:
-            None
-
-        """
-
-        return True
+        self.cmdline = cmdline
+        self.flavor = flavor
 
 
 class System(object):
@@ -86,11 +68,6 @@ class System(object):
     """Class:  System
 
     Description:  Mock of the gen_class.System class.
-
-    Super-Class:  object
-
-    Sub-Classes:
-        None
 
     Methods:
         __init__ -> Class instance initilization.
@@ -138,10 +115,6 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:  None
-
     Methods:
         setUp -> Unit testing initilization.
         test_programlock_fail -> Test ProgramLock fails to lock.
@@ -156,13 +129,13 @@ class UnitTest(unittest.TestCase):
         Description:  Initialization for unit testing.
 
         Arguments:
-            None
 
         """
 
         cfg = collections.namedtuple("Cfg", "memory_threshold")
         self.cfg = cfg(100)
         self.args = {"-c": "config_file", "-d": "config_dir"}
+        self.proglock = ProgramLock(["cmdline"], "FlavorID")
 
     @mock.patch("server_usage.gen_class.ProgramLock")
     @mock.patch("server_usage.gen_libs.load_module")
@@ -174,7 +147,6 @@ class UnitTest(unittest.TestCase):
         Description:  Test ProgramLock fails to lock.
 
         Arguments:
-            None
 
         """
 
@@ -199,11 +171,10 @@ class UnitTest(unittest.TestCase):
         Description:  Test run_program function.
 
         Arguments:
-            None
 
         """
 
-        mock_class.ProgramLock.return_value = ProgramLock([], "Server_Name")
+        mock_class.ProgramLock.return_value = self.proglock
         mock_class.System.return_value = System()
         mock_load.return_value = self.cfg
         mock_info.return_value = {"servername": "Server_Name",