+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [1.0.2] - 2020-06-25
+### Fixed
+- run_program:  Fixed handling command line arguments from SonarQube scan finding.
+- main:  Fixed handling command line arguments from SonarQube scan finding.
+
+### Changed
+- configuration.py.TEMPLATE:  Changed format of configuration file for Mongo connection.
+- Documentation updates.
+
+
 ## [1.0.1] - 2019-05-17
 ### Fixed
 - run_program:  Fixed problem with mutable default arguments issue.